+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### FEATURES
+
+- [ibc-relayer]
+  - Add support for event based channel relaying ([#822])
+  - Graceful handling of packet events in the presence of multiple relayers ([#983])
+
 ### IMPROVEMENTS
 
 - [ibc]
@@ -17,9 +23,11 @@
 - [ibc-relayer-cli]
   - Promote `start-multi` command to `start` ([#911])
 
+[#822]: https://github.com/informalsystems/ibc-rs/issues/822
 [#871]: https://github.com/informalsystems/ibc-rs/issues/871
 [#911]: https://github.com/informalsystems/ibc-rs/issues/911
 [#972]: https://github.com/informalsystems/ibc-rs/issues/972
+[#983]: https://github.com/informalsystems/ibc-rs/issues/983
 
 ## v0.3.2
 *May 21st, 2021*